+++ packages/@expo/cli/src/start/server/metro/createFileMap-fork.ts
@@ -96,7 +96,7 @@ export default function createFileMap(config: ConfigT, options?: CreateFileMapOp
     resetCache: config.resetCache,
     rootDir: projectRoot,
     roots: config.watchFolders,
-    useWatchman: config.resolver.useWatchman ?? false,
+    useWatchman: config.resolver.useWatchman,
     watch,
     watchmanDeferStates: config.watcher.watchman.deferStates,
     // NOTE: (@expo/metro-file-map fork) New option is required for `scopeFallback: true` checks